+++ attributes/git_install.rb
@@ -1,0 +1,1 @@
+default["sls_utils"]["git_install_method"] = "package"