+++ recipes/git_install.rb
@@ -1,0 +1,6 @@
+case node["sls_utils"]["git_install_method"]
+when "source"
+    include_recipe "git::source"
+else
+    include_recipe "git"
+end